Update config to be compatible with v4.2.0 provider (#54)
user_data attribute is no longer "in-place" change, followed by a
stop/start of the instance. This breaks our flow for cfssl and etcd
instances - as ignition will only run on the first boot.  Subsequent
changes and reboot of the instance will ignore new ignition config.

As a workaround I'm moving user_data attribute to be set by
launch_template. Changing launch_template as part of the instance will
cause resource replace:
https://github.com/hashicorp/terraform-provider-aws/blob/main/internal/service/ec2/instance.go#L320

And changing the name of the launch_template also causes resource
replace:
https://github.com/hashicorp/terraform-provider-aws/blob/main/internal/service/ec2/launch_template.go#L41

So if we include the sha1sum of the userdata string in launch_template
name - it will force our instance to be replaced. This works with
current version of the provider but can obviously change.

Other changes:

- "tags" list is deprecated in favour of individual "tag" blocks.
- "aws_s3_bucket_object" deprecated in favour of "aws_s3_object"
- s3 bucket "server_side_encryption_configuration" is no longer
  attribute block, but its own resource