Move to aws_launch_template from aws_launch_configuration
AWS is decomissioning launch configurations:
https://aws.amazon.com/blogs/compute/amazon-ec2-auto-scaling-will-no-longer-add-support-for-new-ec2-features-to-launch-configurations/

They have a timeline for when they disallow creation of new launch
templates, while this might not affect us soon, it makes sense to run a
supported resource if we need to provision anew in a new account.

Also remove the spot price bid variable, not used.