Remove hash suffix from masters and worker s3 location
Use a static location to store ignition under s3 for master and worker nodes.
This will mean that the launch_templates will not update on ignition changes,
but nodes will be able to fetch ignition updates on reboots (if first_boot
flag is set). The goal is to be able to use an operator like kured in order to
perform rolling updates.